groups: remove old update notice and add in new desktop update component
